refactor: split production buttons, train base, debug panel, and CtTopBar (#3878)
Extract library parts for allocation step-button chrome, train dialog
shared state, debug console panel state, and CtTopBar row layout; extend
the debug-console scrim source-contract test to scan declared parts.

Refs #3878
