+++ app/lib/features/game/flame/overlays/debug_console_overlay_panel.dart
@@ -9,6 +9,8 @@ import 'package:flutter/material.dart';
 
 import 'debug_console_controller.dart';
 
+part 'debug_console_overlay_panel_state.dart';
+
 /// `SYS20001` Debug console overlay panel.
 ///
 /// Visual chrome resolves through [EditorialMonoclePalette] tokens — no
@@ -53,136 +55,4 @@ class DebugConsoleOverlayPanel extends StatefulWidget {
   @override
   State<DebugConsoleOverlayPanel> createState() =>
       _DebugConsoleOverlayPanelState();
-}
-
-class _DebugConsoleOverlayPanelState extends State<DebugConsoleOverlayPanel> {
-  late final DebugConsoleController _controller;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = DebugConsoleController(
-      bus: widget.bus,
-      humanPlayerId: widget.humanPlayerId,
-      readOnlyContextProvider: widget.readOnlyContextProvider,
-      onClose: widget.onClose,
-    );
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  void _submit() {
-    setState(() {
-      _controller.submit();
-    });
-  }
-
-  KeyEventResult _handleKey(FocusNode node, KeyEvent event) {
-    final handled = _controller.handleKeyEvent(event);
-    return handled ? KeyEventResult.handled : KeyEventResult.ignored;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = appL10n(context);
-    return Material(
-      color: EditorialMonoclePalette.bgDeep.withValues(
-        alpha: DebugConsoleOverlayPanel.panelBackgroundAlpha,
-      ),
-      borderRadius: BorderRadius.circular(CtRadius.large),
-      child: SizedBox(
-        width: 420,
-        height: 220,
-        child: Padding(
-          padding: const EdgeInsets.all(CtSpacing.m),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: [
-              _buildHeader(l10n),
-              Expanded(
-                child: Container(
-                  padding: const EdgeInsets.all(CtSpacing.s),
-                  color: EditorialMonoclePalette.dialogScrim,
-                  child: _buildLogList(),
-                ),
-              ),
-              const SizedBox(height: CtSpacing.s),
-              _buildInput(l10n),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildHeader(AppLocalizations l10n) {
-    return Row(
-      children: [
-        Expanded(
-          child: Text(
-            l10n.debugConsole_title,
-            style: TextStyle(
-              color: EditorialMonoclePalette.fg,
-              fontWeight: FontWeight.w700,
-            ),
-          ),
-        ),
-        CtIconAction(
-          key: DebugConsoleOverlayPanel.closeButtonKey,
-          tooltip: 'Close debug console',
-          icon: Icons.close,
-          iconColor: EditorialMonoclePalette.fg,
-          onPressed: widget.onClose,
-        ),
-      ],
-    );
-  }
-
-  Widget _buildLogList() {
-    return ListView(
-      children: _controller.lines
-          .map(
-            (line) => Text(
-              line,
-              style: TextStyle(
-                color: EditorialMonoclePalette.fg,
-                fontFamily: 'monospace',
-                fontSize: 12,
-              ),
-            ),
-          )
-          .toList(growable: false),
-    );
-  }
-
-  Widget _buildInput(AppLocalizations l10n) {
-    return Focus(
-      onKeyEvent: _handleKey,
-      child: TextField(
-        key: const ValueKey<String>('debug-console-input'),
-        focusNode: _controller.focusNode,
-        controller: _controller.textController,
-        onSubmitted: (_) => _submit(),
-        style: TextStyle(color: EditorialMonoclePalette.fg),
-        decoration: InputDecoration(
-          isDense: true,
-          hintText: l10n.debugConsole_hintSpawnCivilian,
-          hintStyle: TextStyle(
-            color: EditorialMonoclePalette.muted.withValues(
-              alpha: DebugConsoleOverlayPanel.hintTextAlpha,
-            ),
-          ),
-          filled: true,
-          fillColor: EditorialMonoclePalette.dialogScrim,
-          border: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(CtRadius.medium),
-          ),
-        ),
-      ),
-    );
-  }
 }
