+++ app/lib/features/game/widgets/production/production_allocation_row_buttons.dart
@@ -8,6 +8,10 @@ import '../../../../widgets/ct_gradients.dart';
 import '../../../../widgets/strict_asset_icon.dart';
 import 'production_allocation_repeat_timing.dart';
 
+part 'production_allocation_row_buttons_surface.dart';
+part 'production_allocation_row_buttons_step.dart';
+part 'production_allocation_row_buttons_action.dart';
+
 /// Fixed tappable surface size for the dark editorial-monocle step buttons
 /// per `SPEC/ui/production-panel.md` § Allocation step buttons. The leading
 /// icon keeps its existing ~14–16 px size centered inside this surface.
@@ -18,201 +22,7 @@ const double kProductionAllocationStepButtonSize = 26;
 /// "disabled at 0.3 opacity").
 const double kProductionAllocationStepButtonDisabledOpacity = 0.3;
 
-/// Wraps an icon child in the dark editorial-monocle step-button chrome
-/// (26 × 26 surface with [CtGradients.buttonGradient] inside a 1 px
-/// `EditorialMonoclePalette.border` outline) and fades the entire surface
-/// to [kProductionAllocationStepButtonDisabledOpacity] when disabled.
-///
-/// Shared by the Allocation subpanel's per-recipe ± / maximize / clear
-/// controls **and** the Available subpanel's per-tier labour ± controls,
-/// per `SPEC/ui/production-panel.md` § Allocation step buttons (R51 —
-/// "Enabled tap / long-press semantics and the per-tier production tier
-/// labour controls reuse the same step-button surface contract").
-class ProductionStepButtonSurface extends StatelessWidget {
-  const ProductionStepButtonSurface({
-    super.key,
-    required this.enabled,
-    required this.iconAssetPath,
-    required this.iconSize,
-  });
-
-  final bool enabled;
-  final String iconAssetPath;
-  final double iconSize;
-
-  @override
-  Widget build(BuildContext context) {
-    return Opacity(
-      opacity: enabled ? 1.0 : kProductionAllocationStepButtonDisabledOpacity,
-      child: SizedBox(
-        width: kProductionAllocationStepButtonSize,
-        height: kProductionAllocationStepButtonSize,
-        child: DecoratedBox(
-          decoration: BoxDecoration(
-            gradient: CtGradients.buttonGradient,
-            border: Border.all(
-              color: EditorialMonoclePalette.border,
-              width: 1,
-            ),
-          ),
-          child: Center(
-            child: StrictAssetIcon(
-              assetPath: iconAssetPath,
-              width: iconSize,
-              height: iconSize,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
 typedef ProductionDesiredMapReader = Map<String, int> Function();
 
 /// Returns **true** if the map changed (repeat may continue).
 typedef ProductionAllocationTryStep = bool Function(Map<String, int> current);
-
-/// **+** or **−** with tap and long-press repeat per SPEC/ui/production-panel.md.
-///
-/// Uses [GestureDetector] long-press recognition (same ~500 ms threshold as
-/// [kProductionAllocationRepeatInitialDelay]) plus [Timer.periodic] at
-/// [kProductionAllocationRepeatInterval] for repeats — survives parent rebuilds
-/// from allocation updates better than raw [Listener] timers.
-class ProductionAllocationStepButton extends StatefulWidget {
-  const ProductionAllocationStepButton({
-    super.key,
-    required this.enabled,
-    required this.readDesired,
-    required this.tryStepFromCurrent,
-    required this.semanticLabel,
-    required this.tooltip,
-    required this.assetFileName,
-    this.iconSize = 15,
-  });
-
-  final bool enabled;
-  final ProductionDesiredMapReader readDesired;
-  final ProductionAllocationTryStep tryStepFromCurrent;
-  final String semanticLabel;
-  final String tooltip;
-  final String assetFileName;
-  final double iconSize;
-
-  @override
-  State<ProductionAllocationStepButton> createState() =>
-      _ProductionAllocationStepButtonState();
-}
-
-class _ProductionAllocationStepButtonState
-    extends State<ProductionAllocationStepButton> {
-  Timer? _repeatTimer;
-
-  @override
-  void dispose() {
-    _stopRepeat();
-    super.dispose();
-  }
-
-  void _stopRepeat() {
-    _repeatTimer?.cancel();
-    _repeatTimer = null;
-  }
-
-  void _onLongPress() {
-    if (!widget.enabled) {
-      return;
-    }
-    if (!widget.tryStepFromCurrent(widget.readDesired())) {
-      return;
-    }
-    _repeatTimer?.cancel();
-    _repeatTimer = Timer.periodic(kProductionAllocationRepeatInterval, (_) {
-      if (!mounted) {
-        return;
-      }
-      if (!widget.tryStepFromCurrent(widget.readDesired())) {
-        _stopRepeat();
-      }
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final path = '$kAppIconAssetPrefix${widget.assetFileName}';
-    final surface = ProductionStepButtonSurface(
-      enabled: widget.enabled,
-      iconAssetPath: path,
-      iconSize: widget.iconSize,
-    );
-
-    return Semantics(
-      button: true,
-      enabled: widget.enabled,
-      label: widget.semanticLabel,
-      child: Tooltip(
-        message: widget.tooltip,
-        child: GestureDetector(
-          behavior: HitTestBehavior.opaque,
-          onTap: widget.enabled
-              ? () {
-                  widget.tryStepFromCurrent(widget.readDesired());
-                }
-              : null,
-          onLongPress: widget.enabled ? _onLongPress : null,
-          onLongPressEnd: (_) => _stopRepeat(),
-          onLongPressCancel: () => _stopRepeat(),
-          child: surface,
-        ),
-      ),
-    );
-  }
-}
-
-/// Single tap (maximize / clear) icon control.
-class ProductionAllocationActionIconButton extends StatelessWidget {
-  const ProductionAllocationActionIconButton({
-    super.key,
-    required this.enabled,
-    required this.readDesired,
-    required this.onPressedFromCurrent,
-    required this.semanticLabel,
-    required this.tooltip,
-    required this.assetFileName,
-    this.iconSize = 15,
-  });
-
-  final bool enabled;
-  final ProductionDesiredMapReader readDesired;
-  final void Function(Map<String, int> current) onPressedFromCurrent;
-  final String semanticLabel;
-  final String tooltip;
-  final String assetFileName;
-  final double iconSize;
-
-  @override
-  Widget build(BuildContext context) {
-    final path = '$kAppIconAssetPrefix$assetFileName';
-    final surface = ProductionStepButtonSurface(
-      enabled: enabled,
-      iconAssetPath: path,
-      iconSize: iconSize,
-    );
-
-    return Semantics(
-      button: true,
-      enabled: enabled,
-      label: semanticLabel,
-      child: Tooltip(
-        message: tooltip,
-        child: Material(
-          color: Colors.transparent,
-          child: InkWell(
-            onTap: enabled ? () => onPressedFromCurrent(readDesired()) : null,
-            child: surface,
-          ),
-        ),
-      ),
-    );
-  }
-}
